requirements.txt
django
djangorestframework
djangorestframework-simplejwt
requests
cryptography
django-environ
dotenv
mozilla-django-oidc
opensearch-py